requirements.txt
-e .

aws-cdk.aws-lambda
aws_cdk.aws_iam